Don't delete symbolic links if they point to a directory
If a symlink was pointing to a directory (e.g. if .vim was a
symlink), it would get removed even though we never created it.

Fixes issue #2

+++ nix/bootstrap.sh
@@ -43,7 +43,9 @@ for i in ~/.*; do
     [ ! -h "$i" ] && continue
 
     # We have a link: Is it stale? If so, delete it ...
-    if [ ! -f "$i" ]; then
+    # Since we can't use readlink, assume that if the link is
+    # not pointing to a file or a directory that it is stale.
+    if [ ! -f "$i" -a ! -d "$i" ]; then
         echo "Deleting stale link: $i"
         rm "$i"
     fi
